Move a few design specific methods off of block
