+++ armi/reactor/blocks.py
@@ -71,8 +71,6 @@ class Block(composites.Composite):
     Blocks are stacked together to form assemblies.
     """
 
-    # nuclides that will be put in A.NIP3 but not A.bURN (these will not deplete!)
-    inerts = []
     uniqID = 0
 
     # dimension used to determine which component defines the block's pitch
@@ -109,15 +107,8 @@ class Block(composites.Composite):
             self.spatialLocator = grids.IndexLocation(0, 0, k, None)
         self.p.orientation = numpy.array((0.0, 0.0, 0.0))
 
-        self.nuclides = (
-            []
-        )  # TODO: list of nuclides present in this block (why not just density.keys()?)
         self.points = []
         self.macros = None
-
-        self.numLfpLast = {}  # for FGremoval
-        self.history = []  # memory of shuffle locations
-        self.lastkInf = 0.0  # for tracking k-inf vs. time slope.
 
         # flag to indicated when DerivedShape children must be updated.
         self.derivedMustUpdate = False
@@ -686,18 +677,6 @@ class Block(composites.Composite):
 
         return 4.0 * self.getFlowAreaPerPin() / self.getWettedPerimeter()
 
-    def getCladdingOR(self):
-        clad = self.getComponent(Flags.CLAD)
-        return clad.getDimension("od") / 2.0
-
-    def getCladdingIR(self):
-        clad = self.getComponent(Flags.CLAD)
-        return clad.getDimension("id") / 2.0
-
-    def getFuelRadius(self):
-        fuel = self.getComponent(Flags.FUEL)
-        return fuel.getDimension("od") / 2.0
-
     def adjustUEnrich(self, newEnrich):
         """
         Adjust U-235/U-238 mass ratio to a mass enrichment
@@ -723,95 +702,6 @@ class Block(composites.Composite):
                 self.setNumberDensity("U238", tU * (1.0 - newEnrich))
 
         self.completeInitialLoading()
-
-    def adjustSmearDensity(self, value, bolBlock=None):
-        r"""
-        modifies the *cold* smear density of a fuel pin by adding or removing fuel dimension.
-
-        Adjusts fuel dimension while keeping cladding ID constant
-
-        sd = fuel_r**2/clad_ir**2  =(fuel_od/2)**2 / (clad_id/2)**2 = fuel_od**2 / clad_id**2
-        new fuel_od = sqrt(sd*clad_id**2)
-
-        useful for optimization cases
-
-        Parameters
-        ----------
-
-        value : float
-            new smear density as a fraction.  This fraction must
-            evaluate between 0.0 and 1.0
-
-        bolBlock : Block, optional
-            See completeInitialLoading. Required for ECPT cases
-
-        """
-        if 0.0 >= value or value > 1.0:
-            raise ValueError(
-                "Cannot modify smear density of {0} to {1}. Must be a positive fraction"
-                "".format(self, value)
-            )
-        fuel = self.getComponent(Flags.FUEL)
-        if not fuel:
-            runLog.warning(
-                "Cannot modify smear density of {0} because it is not fuel".format(
-                    self
-                ),
-                single=True,
-                label="adjust smear density",
-            )
-            return
-
-        clad = self.getComponent(Flags.CLAD)
-        cladID = clad.getDimension("id", cold=True)
-        fuelID = fuel.getDimension("id", cold=True)
-
-        if fuelID > 0.0:  # Annular fuel (Adjust fuel ID to get new smear density)
-            fuelOD = fuel.getDimension("od", cold=True)
-            newID = fuelOD * math.sqrt(1.0 - value)
-            fuel.setDimension("id", newID)
-        else:  # Slug fuel (Adjust fuel OD to get new smear density)
-            newOD = math.sqrt(value * cladID ** 2)
-            fuel.setDimension("od", newOD)
-
-        # update things like hm at BOC and smear density parameters.
-        self.completeInitialLoading(bolBlock=bolBlock)
-
-    def adjustCladThicknessByOD(self, value):
-        """Modifies the cladding thickness by adjusting the cladding outer diameter."""
-        clad = self._getCladdingComponentToModify(value)
-        if clad is None:
-            return
-        innerDiam = clad.getDimension("id", cold=True)
-        clad.setDimension("od", innerDiam + 2.0 * value)
-
-    def adjustCladThicknessByID(self, value):
-        """
-        Modifies the cladding thickness by adjusting the cladding inner diameter.
-
-        Notes
-        -----
-        This WILL adjust the fuel smear density
-        """
-        clad = self._getCladdingComponentToModify(value)
-        if clad is None:
-            return
-        od = clad.getDimension("od", cold=True)
-        clad.setDimension("id", od - 2.0 * value)
-
-    def _getCladdingComponentToModify(self, value):
-        clad = self.getComponent(Flags.CLAD)
-        if not clad:
-            runLog.warning(
-                "{} does not have a cladding component to modify.".format(self)
-            )
-        if value < 0.0:
-            raise ValueError(
-                "Cannot modify {} on {} due to a negative modifier {}".format(
-                    clad, self, value
-                )
-            )
-        return clad
 
     def getLocation(self):
         """Return a string representation of the location."""
